+++ _gtm_head_script_html.d.ts
@@ -1,0 +1,6 @@
+import type { _gtm_script_html_opts_type } from './_gtm_script_html_opts_type';
+/**
+ * Google Tag Manager script html to place at the top of `<head>`
+ */
+export declare function _gtm_head_script_html(opts?: _gtm_script_html_opts_type): string;
+export { _gtm_head_script_html as _html__script__head__gtm };