+++ _gtm_script_html.d.ts
@@ -1,0 +1,9 @@
+import type { _gtm_script_html_opts_type } from './_gtm_script_html_opts_type';
+/**
+ * Html to add gtm.js to the page
+ * @param opts
+ * @param opts.GTM_ID
+ * @returns {string}
+ */
+export declare function _gtm_script_html(opts?: _gtm_script_html_opts_type): string;
+export { _gtm_script_html as _html__script__gtm, _gtm_script_html as _script__gtm, };