+++ _gtm_script_html_opts_type.ts
@@ -1,0 +1,5 @@
+import type { error_ctx_type } from '@ctx-core/error'
+export interface _gtm_script_html_opts_type extends error_ctx_type {
+	GTM_ID?:string
+	dataLayer?:[]
+}