+++ gtm_body_script_html_opts_type.ts
@@ -1,0 +1,6 @@
+import type { error_ctx_type } from '@ctx-core/error'
+export interface gtm_body_script_html_opts_type extends error_ctx_type {
+	GTM_ID?:string
+	GOOGLE_TRACKING_ID?:string
+	GA_ID?:string
+}