+++ init_dataLayer.d.ts
@@ -1,0 +1,2 @@
+export declare function init_dataLayer(): void;
+export { init_dataLayer as init__dataLayer };