+++ push_dataLayer.d.ts
@@ -1,0 +1,2 @@
+export declare function push_dataLayer(obj: any, ...arg_a1: unknown[]): void;
+export { push_dataLayer as push__dataLayer };